Rewrite POST servercheck to GO (#3935)
* Rewrite POST serverchecks to GO

* License headers missing on some go files

* Address PR review

* Fix to_extension active type

* Update serverchecks_test.go

* Fix db_helper for ats

* Fix tests

+++ traffic_ops/traffic_ops_golang/server/servers_assignment.go
@@ -62,7 +62,7 @@ func AssignDeliveryServicesToServerHandler(w http.ResponseWriter, r *http.Reques
 		api.HandleErr(w, r, inf.Tx.Tx, http.StatusBadRequest, err, nil)
 		return
 	}
-	serverName, ok, err := dbhelpers.GetServerNameFromID(inf.Tx.Tx, int64(server))
+	serverName, ok, err := dbhelpers.GetServerNameFromID(inf.Tx.Tx, server)
 	if err != nil {
 		api.HandleErr(w, r, inf.Tx.Tx, http.StatusInternalServerError, nil, errors.New("getting server name from ID: "+err.Error()))
 		return
